Move Row to cursor.rs
